// src/app/Enums/enums.ts
export enum Marks {
    VeryLow = -2,
    Low = -1,
    Default = 0,
    High = 1,
    VeryHigh = 2,
    Unvoted = -246111
}

export enum Colors {
    VeryLow = '#FF0000',//'#ae0000',
    Low = '#F08080',//'#f18c8c',
    Default = '#FFFF00',//'#ffe81a',
    High = '#ADFF2F',//'rgb(177, 214, 104)',
    VeryHigh = '#00FF00'//'#228B22',
}

export enum MarkTypes {
    BadGood = 1,
    ProCon = 2
}
// background: #ae0000;
// background: #f18c8c;
// background: #ffe81a;
// background: rgb(177, 214, 104);
// background: #228B22;
// e4eff1fb